refactor: Change the order of the functions
This change brings the view more button code on top of
the view raw data button code to make debugging the two easier.

The display of the view more button depends on the behaviour
of the view raw data button.

The logic of the view raw data button was at the top of the file
while the logic of the view more button was at the bottom of
the file. This obliged the developer to scroll through all the file
up and down to test changes.