refactor: Make view_more() more efficient
view_more() returns an html table by reading a csv string
representing the filtered city data (generated by load_data())

Before this change, each time view_more() was called, it would
create a new table with five new rows by reading more lines
starting at the beginning of the csv string.

If it's called once, it reads the csv string from the
first to the 5th line. If it's called twice, it reads
the csv string from the first to the 10th line...

The problem with that implementation was that each call
made the function slower.

The current implementation makes view_more() receive a
table that it generated before with an anterior call
as an input.
It reads only 5 lines from the csv string and it skips the lines
that were used to generate the previous table. Then, it appends the
new rows instead of creating a new table.

The first call makes the function read 5 lines starting at
the beginning of the string.
The second call makes the function read 5 lines starting at the 6th
line (inclusive) of the csv string.
The third call makes the function read 5 lines starting
at the 11th(inclusive) line...